frontend/scripts/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
PyJWT==2.8.0
Werkzeug==2.3.7
